follow ruby style guid
